ml/requirements.txt
xgboost==3.2.0
fastapi==0.135.3
uvicorn[standard]==0.44.0
supabase==2.28.3
scikit-learn==1.8.0
pandas==3.0.2
numpy==2.4.4
python-dotenv==1.2.2
pydantic==2.13.0
coverage==7.13.5
